ml/requirements.txt
pandas==2.1.1
scikit-learn==1.3.1
xgboost==2.0.0
numpy==1.26.0
matplotlib==3.8.0
seaborn==0.13.0
joblib==1.3.2
openai==1.2.0
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
pytest==7.4.2
celery==5.3.4
redis==5.0.1
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
tqdm==4.66.1
